refactor: page별로 파일 분리 - ProjectList - 프로젝트 리스트 페이지 파일 분리
